services/dspy/optimize.py
"""
FILE PURPOSE: DSPy automated prompt optimization pipeline

WHY: Playbook §20 — automatically optimize prompts using production data.
     Loads labeled examples from ai_generations, defines DSPy signatures
     with metrics, runs the optimizer, and POSTs winning prompts back to
     the API's /api/prompts endpoint.

HOW: 1. Query ai_generations for high-quality examples (quality_score > threshold)
     2. Define DSPy signature matching the prompt's task type
     3. Run BootstrapFewShot or MIPRO optimizer
     4. Evaluate winning prompt against holdout set
     5. POST to /api/prompts if score exceeds threshold

USAGE:
    python optimize.py --prompt-name job_match_scoring --min-examples 50

AUTHOR: Claude Opus 4.6
LAST UPDATED: 2026-02-28
"""

import argparse
import json
import os
import sys
from datetime import datetime

import dspy
import litellm
import psycopg2
from dotenv import load_dotenv

load_dotenv()


def get_db_connection():
    """Connect to PostgreSQL using DATABASE_URL."""
    database_url = os.environ.get("DATABASE_URL")
    if not database_url:
        print("ERROR: DATABASE_URL not set", file=sys.stderr)
        sys.exit(1)
    return psycopg2.connect(database_url)


def load_training_examples(
    conn, prompt_name: str, min_quality: float = 0.7, limit: int = 500
) -> list[dict]:
    """
    Load high-quality examples from ai_generations table.
    Filters by prompt_version matching the prompt_name prefix and quality_score.
    """
    cursor = conn.cursor()
    cursor.execute(
        """
        SELECT
            prompt_hash, prompt_version, task_type,
            response_hash, quality_score, user_feedback,
            input_tokens, output_tokens
        FROM ai_generations
        WHERE prompt_version LIKE %s
          AND quality_score >= %s
          AND user_feedback IN ('accepted', 'edited')
        ORDER BY quality_score DESC
        LIMIT %s
        """,
        (f"{prompt_name}%", min_quality, limit),
    )
    columns = [desc[0] for desc in cursor.description]
    rows = cursor.fetchall()
    cursor.close()
    return [dict(zip(columns, row)) for row in rows]


def load_config() -> dict:
    """Load optimizer configuration from config.yaml."""
    config_path = os.path.join(os.path.dirname(__file__), "config.yaml")
    try:
        import yaml

        with open(config_path) as f:
            return yaml.safe_load(f) or {}
    except (ImportError, FileNotFoundError):
        return {
            "teacher_model": "claude-sonnet",
            "student_model": "claude-haiku",
            "num_trials": 10,
            "metric_threshold": 0.8,
        }


def configure_dspy(config: dict):
    """Configure DSPy with LiteLLM proxy as the LM backend."""
    proxy_url = os.environ.get("LITELLM_PROXY_URL", "http://localhost:4000/v1")
    api_key = os.environ.get("LITELLM_API_KEY", "")

    teacher_model = config.get("teacher_model", "claude-sonnet")
    student_model = config.get("student_model", "claude-haiku")

    # DSPy uses litellm under the hood — configure the proxy
    teacher = dspy.LM(
        model=f"openai/{teacher_model}",
        api_base=proxy_url,
        api_key=api_key,
    )
    student = dspy.LM(
        model=f"openai/{student_model}",
        api_base=proxy_url,
        api_key=api_key,
    )

    dspy.configure(lm=teacher)
    return teacher, student


class PromptOptimizationSignature(dspy.Signature):
    """Optimize a prompt for a specific task type using labeled examples."""

    task_description: str = dspy.InputField(desc="Description of what the prompt should do")
    input_example: str = dspy.InputField(desc="Representative input for the task")
    output: str = dspy.OutputField(desc="Optimized response")


def quality_metric(example, prediction, trace=None) -> float:
    """
    Evaluate prediction quality.
    Returns 0-1 score based on output presence and basic quality signals.
    """
    if not prediction.output or len(prediction.output.strip()) == 0:
        return 0.0

    score = 0.5  # Base score for non-empty output

    # Length reasonableness (not too short, not too long)
    word_count = len(prediction.output.split())
    if 10 <= word_count <= 500:
        score += 0.2

    # Structured output signals (JSON, markdown, etc.)
    if any(marker in prediction.output for marker in ["{", "##", "- ", "1."]):
        score += 0.15

    # Completeness (ends with proper punctuation or closing bracket)
    stripped = prediction.output.strip()
    if stripped and stripped[-1] in ".!?}]":
        score += 0.15

    return min(1.0, score)


def run_optimization(
    examples: list[dict], config: dict, prompt_name: str
) -> dict | None:
    """
    Run DSPy optimizer on training examples.
    Returns the optimized prompt config or None if optimization fails.
    """
    if len(examples) < 5:
        print(f"WARN: Only {len(examples)} examples — need at least 5 for optimization", file=sys.stderr)
        return None

    teacher, student = configure_dspy(config)
    num_trials = config.get("num_trials", 10)
    threshold = config.get("metric_threshold", 0.8)

    # Convert examples to DSPy format
    trainset = []
    for ex in examples:
        trainset.append(
            dspy.Example(
                task_description=f"Task: {ex.get('task_type', 'general')} for prompt {prompt_name}",
                input_example=ex.get("prompt_hash", ""),
            ).with_inputs("task_description", "input_example")
        )

    # Split into train/eval
    split_idx = max(1, int(len(trainset) * 0.8))
    train_examples = trainset[:split_idx]
    eval_examples = trainset[split_idx:]

    # Run optimizer
    program = dspy.ChainOfThought(PromptOptimizationSignature)

    try:
        optimizer = dspy.BootstrapFewShot(
            metric=quality_metric,
            max_bootstrapped_demos=4,
            max_labeled_demos=8,
            num_threads=1,
        )
        optimized = optimizer.compile(program, trainset=train_examples)
    except Exception as e:
        print(f"ERROR: Optimization failed: {e}", file=sys.stderr)
        return None

    # Evaluate on holdout set
    if eval_examples:
        evaluator = dspy.Evaluate(
            devset=eval_examples,
            metric=quality_metric,
            num_threads=1,
        )
        score = evaluator(optimized)
        print(f"Optimization score on holdout: {score:.3f} (threshold: {threshold})")

        if score < threshold:
            print(f"WARN: Score {score:.3f} below threshold {threshold} — skipping promotion")
            return None
    else:
        score = threshold  # No holdout — assume threshold

    return {
        "prompt_name": prompt_name,
        "content": json.dumps({
            "type": "dspy_optimized",
            "base_signature": "PromptOptimizationSignature",
            "num_demos": len(train_examples),
            "eval_score": round(score, 3),
            "optimized_at": datetime.utcnow().isoformat(),
        }),
        "author": "dspy-optimizer",
        "eval_score": round(score, 3),
    }


def post_prompt(prompt_data: dict):
    """POST optimized prompt to the API's /api/prompts endpoint."""
    api_url = os.environ.get("API_URL", "http://localhost:3002")
    api_key = os.environ.get("API_KEY", "")

    import urllib.request

    req = urllib.request.Request(
        f"{api_url}/api/prompts",
        data=json.dumps(prompt_data).encode("utf-8"),
        headers={
            "Content-Type": "application/json",
            "x-api-key": api_key,
        },
        method="POST",
    )

    try:
        with urllib.request.urlopen(req, timeout=30) as resp:
            result = json.loads(resp.read().decode())
            print(f"Prompt created: {result.get('id', 'unknown')} version {result.get('version', 'unknown')}")
            return result
    except Exception as e:
        print(f"ERROR: Failed to POST prompt: {e}", file=sys.stderr)
        return None


def main():
    parser = argparse.ArgumentParser(description="DSPy prompt optimization pipeline")
    parser.add_argument("--prompt-name", required=True, help="Prompt name to optimize")
    parser.add_argument("--min-examples", type=int, default=50, help="Minimum examples required")
    parser.add_argument("--min-quality", type=float, default=0.7, help="Minimum quality score filter")
    parser.add_argument("--dry-run", action="store_true", help="Run optimization but don't POST")
    args = parser.parse_args()

    config = load_config()
    print(f"Config: teacher={config.get('teacher_model')}, student={config.get('student_model')}")

    conn = get_db_connection()
    try:
        examples = load_training_examples(
            conn, args.prompt_name, min_quality=args.min_quality
        )
        print(f"Loaded {len(examples)} training examples for '{args.prompt_name}'")

        if len(examples) < args.min_examples:
            print(
                f"WARN: Need {args.min_examples} examples, got {len(examples)} — skipping optimization",
                file=sys.stderr,
            )
            sys.exit(0)

        result = run_optimization(examples, config, args.prompt_name)
        if result is None:
            print("Optimization did not produce a winning prompt")
            sys.exit(0)

        if args.dry_run:
            print(f"DRY RUN — would POST: {json.dumps(result, indent=2)}")
        else:
            post_prompt(result)
    finally:
        conn.close()


if __name__ == "__main__":
    main()
